python-app/requirements.txt
requests[socks]>=2.31.0
PySocks==1.7.1
python-bitcoinlib>=0.12.0 